+++ src/Sulu/Bundle/ContentBundle/Controller/NodeResourcelocatorController.php
@@ -109,6 +109,6 @@ class NodeResourcelocatorController extends RestController implements ClassResou
      */
     private function getSession()
     {
-        return $this->get('doctrine_phpcr.default_session');
+        return $this->get('doctrine_phpcr.session');
     }
 }